lot of fixes
git-svn-id: https://lmms.svn.sf.net/svnroot/lmms/trunk/lmms@124 0778d3d1-df1d-0410-868b-ea421aaaa00d

+++ include/track_container.h
@@ -31,9 +31,9 @@
 
 #ifdef QT4
 
-#include <QScrollArea>
-#include <QVector>
-#include <QMainWindow>
+#include <QtGui/QScrollArea>
+#include <QtCore/QVector>
+#include <QtGui/QMainWindow>
 
 #else
 
@@ -89,7 +89,7 @@ public:
 		return( FALSE );
 	}
 
-	unsigned int FASTCALL countTracks( track::trackTypes _tt =
+	Uint16 FASTCALL countTracks( track::trackTypes _tt =
 					track::TOTAL_TRACK_TYPES ) const;
 
 	void FASTCALL setMutedOfAllTracks( bool _muted );
